+++ CHANGELOG.md
@@ -114,6 +114,7 @@ Whitespace conventions:
 - Fixed `Integer("0")` raising `ArgumentError` instead of parsing as 0
 - Fixed `JSON#parse` to raise `JSON::ParserError` for invalid input
 - `Module#append_features` now detects cyclic includes
+- `Process.clock_gettime(Process::CLOCK_MONOTONIC)` will now return true monotonic values or raise `Errno::EINVAL` if no monotonic clock is available